+++ packages/terra-alert/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 * Added
-  * Added visually hidden commas between alert content.
+  * Added visually hidden default title for screen readers to read with custom titles.
 
 ## 4.80.0 - (September 21, 2023)